backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv==1.0.1
motor==3.3.1
pydantic==2.6.4
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
